// src/app/models/enums.model.ts
export enum NiveauEtude {
  CFEE = 'CFEE',
  BFEM = 'BFEM',
  BAC = 'BAC',
  LICENCE = 'LICENCE',
  MASTER = 'MASTER',
  DOCTORAT = 'DOCTORAT'
}

export enum TypeContrat {
  CDI = 'CDI',
  CDD = 'CDD',
  STAGE = 'STAGE',
  FREELANCE = 'FREELANCE',
  ALTERNANCE = 'ALTERNANCE',
  THESE = 'THESE'
}

export enum Statut {
  ACTIVE = 'ACTIVE',
  EXPIREE = 'EXPIREE',
  POURVUE = 'POURVUE'
}

export enum StatutCandidature {
  EN_ATTENTE_VALIDATION_ADMIN = 'EN_ATTENTE_VALIDATION_ADMIN',
  REJETEE_ADMIN = 'REJETEE_ADMIN',
  VALIDEE_ADMIN = 'VALIDEE_ADMIN',
  VUE_RECRUTEUR = 'VUE_RECRUTEUR',
  ACCEPTEE_RECRUTEUR = 'ACCEPTEE_RECRUTEUR',
  REFUSEE_RECRUTEUR = 'REFUSEE_RECRUTEUR'
}